chore: api 모듈 rename (api -> api-fcfs)
선착순 처리와 쿠폰 처리 application 분리 위함
