+++ fcfs-coupon-api-fcfs/src/main/java/com/coupop/fcfscoupon/api/fcfs/dto/IssuanceRequest.java
@@ -1,4 +1,4 @@
-package com.coupop.fcfscoupon.api.dto;
+package com.coupop.fcfscoupon.api.fcfs.dto;
 
 import com.coupop.fcfscoupon.common.inputvalidation.EmailValidation;
 import jakarta.validation.constraints.Email;
