+++ fcfs-coupon-api-fcfs/src/main/java/com/coupop/fcfscoupon/api/fcfs/support/RequestTime.java
@@ -1,4 +1,4 @@
-package com.coupop.fcfscoupon.api.support;
+package com.coupop.fcfscoupon.api.fcfs.support;
 
 import java.time.LocalTime;
 import org.springframework.context.annotation.Scope;
