+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 rootProject.name = "fcfs-coupon"
-include("fcfs-coupon-api")
+include("fcfs-coupon-api-fcfs")
 include("fcfs-coupon-mq")
 include("fcfs-coupon-domain-fcfs")
 include("fcfs-coupon-domain-coupon")
